requirements.txt
matplotlib>=3.5.0
pandas>=1.5.0
numpy>=1.21.0
requests>=2.25.0
python-dateutil>=2.8.0
tqdm>=4.60.0